Use wall-time, not CPU time (aka `clock()`).
Profilers usually use clock() to measure performance. I consider this a
mistake. clock() does not measure time in IO or time waiting on
other processes such as databases.

clock() is designed to time CPU intensive work. However, since python is
often used for ‘glue’ code for complex systems, clock() misses the
point. In a web framework, often the majority of the time is spent
executing a db query, and this is valuable information.