Fix Windows filename issue for URLs with a query string
Windows doesn't allow "?" in filenames, which causes the pyinstrument Django middleware to fail for any URL with a query string. 

I've added a simple replace of "?" with "_qs_" to the file path, but I'm not sure it's the right approach; query strings can get pretty long, and the resulting var cuts off at 100 chars (which is reasonable). Another approach might be to write the query string into the HTML output, but that's not such a quick pull request 😄