Allow --load-prev from file if it exists
The previous behaviour of loading from the report
directory is still kept as the default.

However, if the argument is a valid path for an
existing file and NOT a valid identifier, load
from the file instead.

This is handy when reports get copied around
and shared between different people and systems.