Loading and saving pyisession objects from the command line (#189)
* Output a session file using `-r session`

* Neaten logic

* Improve comment

* Allow --load-prev from file if it exists

The previous behaviour of loading from the report
directory is still kept as the default.

However, if the argument is a valid path for an
existing file and NOT a valid identifier, load
from the file instead.

This is handy when reports get copied around
and shared between different people and systems.

* Add test for command-line session saving and loading

* Add a separate option for loading from a file, for future-proofing

* Guess the renderer based on the outfile extension, if given

* Fix overzealous refactor of FrameRenderer

Co-authored-by: Steinar V. Kaldager <steinar.kaldager@oda.com>