Use slightly more recent IPython API.
The new ast transformer are a bit more flexible instead of just
appending/prepending some code and allow full Python constructs.

- you can use ``__code__`` for where the original code should go, and
  ``__ret__``
- by default it can also mangle all names that start with `___` (3
  underscore), if you need some more hygienic templating.
  The names are still leaked in global namespace but with invalid python
  indentifiers names starting with `mangle-` (where `-` is invalid in an
  identifier).

This allow us to later clean the namespace manually in the magic.

And this also allwo us to also return the original value from the cell,
I don't know if this is something desirable, but I added it nonetheless.