Share some options between the CLI and magic.
This refactor compute_render_options, to be shared between CLI and
magic, it changes the last parameter from a file-like output to a
boolean to control wether or not the output will support color and
unicode.

The only thing I don't really like is that we redefine
the help text options.

Co-authored-by: Joe Rickerby <joerick@mac.com>